requirements.txt
django
djangorestframework
python-dotenv
mysql-connector-python
django-filter
django-extensions
drf-yasg